Update docs: add sections for UnifiedPush and subscribing to a topic
As this document file has not been quite clear on what the role actually installs and manages and how a ntfy server supports UnifiedPush together with its own notification system based on subscribing to a topic (the description about it has been lacked mainly because this file was initially copied from the MDAD project), this commit adds descriptions about how UnifiedPush works, how it is related to this role, and how ntfy Android/iOS/Web apps should be configured.

In short, this commit aims to clarify that ntfy supports UnifiedPush but it has its own functions which are not related to UnifiedPush, and describe how those ntfy-specific functions can be configured.

Signed-off-by: Suguru Hirahara <acioustick@noreply.codeberg.org>