Update docs/configuring-ntfy.md: add an example for Matrix server
This example is for those who install a ntfy server with the MASH Ansible playbook and install (or have already installed) a Matrix server with different methods than using the matrix-docker-ansible-deploy Ansible playbook where ntfy role is available as well. It is not likely to often, but one might prefer managing non-Matrix services on the MASH playbook in order to manage only Matrix related services with the MDAD playbook.

Signed-off-by: Suguru Hirahara <acioustick@noreply.codeberg.org>