Update configuring-ntfy.md (#13)
* Update docs/configuring-ntfy.md: emphasize the default status without access control

Signed-off-by: Suguru Hirahara <acioustick@noreply.codeberg.org>

* Update docs/configuring-ntfy.md: add a note about automatic configuration of access control by the playbook

As the role takes care of the step on behalf of the user which the ntfy's official documentation claims needs to be done, it might as well be explained.

Signed-off-by: Suguru Hirahara <luixxiul@users.noreply.github.com>

* Update docs/configuring-ntfy.md: add a recommendation to enable access control

Signed-off-by: Suguru Hirahara <acioustick@noreply.codeberg.org>

* Update docs/configuring-ntfy.md: add an example for Matrix server

This example is for those who install a ntfy server with the MASH Ansible playbook and install (or have already installed) a Matrix server with different methods than using the matrix-docker-ansible-deploy Ansible playbook where ntfy role is available as well. It is not likely to often, but one might prefer managing non-Matrix services on the MASH playbook in order to manage only Matrix related services with the MDAD playbook.

Signed-off-by: Suguru Hirahara <acioustick@noreply.codeberg.org>

* Update docs/configuring-ntfy.md: add an instruction to log in when authentication is enabled

As it is fairly easy to miss that you'd need to log in to your account on the ntfy's Android app or the web app if enabled, this adds the instruction about it.

Signed-off-by: Suguru Hirahara <acioustick@noreply.codeberg.org>

* Fix capitalization: Ntfy → ntfy

Signed-off-by: Suguru Hirahara <acioustick@noreply.codeberg.org>

---------

Signed-off-by: Suguru Hirahara <acioustick@noreply.codeberg.org>
Signed-off-by: Suguru Hirahara <luixxiul@users.noreply.github.com>
Co-authored-by: Suguru Hirahara <acioustick@noreply.codeberg.org>